// apps/web/src/entities/generated/entity-details.gen.ts
export type EntityDetailByEntity = {
  product: z.output<typeof productWithFoodOut>;
  recipe: z.output<typeof recipeOut>;
  ingredient: z.output<typeof ingredientWithFoodOut>;
  location: z.output<typeof infLocation>;
  inventory: z.output<typeof inventoryWithLocationAndProductOut>;
  meal: z.output<typeof mealOut>;
  ledgerParty: z.output<typeof ledgerPartyOut>;
  ledgerTransfer: z.output<typeof ledgerTransferOut>;
  project: z.output<typeof projectOut>;
  task: z.output<typeof taskOut>;
  vendor: z.output<typeof vendorOut>;
  purchase: z.output<typeof purchaseOut>;
  financialAccount: z.output<typeof financialAccountOut>;
  financialTransaction: z.output<typeof financialTransactionOut>;
  wish: z.output<typeof wishOut>;
  expense: z.output<typeof expenseOut>;
};